refactor(6.4): extract hooks and remove props-to-state antipattern
