+++ src/components/lists/cards-list.tsx
@@ -1,15 +1,15 @@
 'use client'
 
-import { useState, useEffect } from 'react'
+import { useState } from 'react'
 import Popup, { CardAction } from '../popups/popup'
 import { ICard } from '@/types/ICard'
 import { IUser } from '@/types/IUser'
-import { CompletionStatus } from '@/types/CompletionStatusEnum'
 import AddCard from '../cards/add-card'
 import { IActivity } from '@/types/IActivity'
 import { useMounted } from '@/lib/helpers/useMounted'
 import { getBorderColor } from '@/lib/helpers/getBorderColor'
 import formatCardName from '@/lib/helpers/formatCardName'
+import { useCardFiltering } from '@/hooks/use-card-filtering'
 import {
   DocumentMinusIcon,
   DocumentPlusIcon,
@@ -28,63 +28,19 @@ export default function CardsList({
   activity?: IActivity
 }) {
   const mounted = useMounted()
+  const { cardsOfTheDay, allActiveCards, allInactiveCards, hash } = useCardFiltering(
+    activity,
+    userDetails,
+    mounted
+  )
 
   const [showModal, setShowModal] = useState(false)
   const [popupItem, setPopupItem] = useState<{ card: ICard; action: CardAction }>(
     {} as { card: ICard; action: CardAction }
   )
-  const [popupUserDetails, setPopupUserDetails] = useState<IUser>({} as IUser)
-  const [cardsOfTheDay, setCardsOfTheDay] = useState<ICard[]>([])
-  const [allActiveCards, setAllActiveCards] = useState<ICard[]>([])
-  const [allInactiveCards, setAllInactiveCards] = useState<ICard[]>([])
-  const [hash, setHash] = useState('')
-
-  useEffect(() => {
-    const cards: any[] | ICard = activity?.cards || []
-
-    if (cards.length) {
-      cards.map((card) => {
-        let num = `${atob(card.cardId).split('-')[1]}`
-        if (num.length === 1) {
-          card.number = `00${num}`
-        } else if (num.length === 2) {
-          card.number = `0${num}`
-        } else {
-          card.number = `${num}`
-        }
-      })
-      cards.sort((a, b) => a.number - b.number)
-      cards.map((card) => delete card.number)
-    }
-
-    const todayCards = cards?.filter((card) => {
-      return (
-        card.completionStatus === CompletionStatus.REVIEW ||
-        (card.completionStatus !== CompletionStatus.INACTIVE &&
-          card.completionStatus !== CompletionStatus.COMPLETED &&
-          (new Date().toLocaleDateString('en-US', { timeZone: 'EST' }) === card.nextShowDate ||
-            new Date().toLocaleDateString('en-US', { timeZone: 'EST' }) === card.addedOn))
-      )
-    })
-    setCardsOfTheDay(todayCards)
-
-    const activeCards = cards?.filter((card) => card.completionStatus !== CompletionStatus.INACTIVE)
-    setAllActiveCards(activeCards)
-
-    const inactiveCards = cards?.filter(
-      (card) => card.completionStatus === CompletionStatus.INACTIVE
-    )
-    setAllInactiveCards(inactiveCards)
-
-    if (mounted) {
-      const hashReducer = window.location.hash
-      setHash(hashReducer)
-    }
-  }, [userDetails, activity, mounted])
 
   function openCardPopup(card: ICard, action: CardAction) {
     setPopupItem({ card, action })
-    setPopupUserDetails(userDetails)
     setShowModal(true)
   }
 
@@ -333,7 +289,7 @@ export default function CardsList({
           config={{
             type: 'manageCard',
             isMain,
-            user: popupUserDetails,
+            user: userDetails,
             activity: activity!,
             item: popupItem,
           }}
